tests/load/requirements.txt
websocket-client>=1.6.0
PyJWT>=2.8.0
